Fader debug star/stop feature
I added a coroutine to be able to loop through the fader effect beginning/end. An editor script was also created so allow for runtime Start/Stop buttons in the Inspector. Speed can be adjusted using the "LoopSpeed" var. One shots were also created in you just want to see it one time.